--- src/main/java/com/qm/tool/core/StringUtils.java
@@ -1,48 +1,0 @@
-package com.qm.tool.core;
-
-/**
- * 字符串工具类
- */
-public class StringUtils {
-
-    /**
-     * 判断字符串是否相等
-     *
-     * @param s1
-     * @param s2
-     * @return
-     */
-    public static boolean equals(String s1, String s2) {
-        if (s1 == null || s2 == null) {
-            return false;
-        }
-        return s1.equals(s2);
-    }
-
-    /**
-     * 判断字符串是否为空
-     *
-     * @param str
-     * @return
-     */
-    public static boolean isEmpty(String str) {
-        if (str == null || "".equals(str.trim())) {
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * 判断字符串是否为空
-     *
-     * @param str
-     * @return
-     */
-    public static boolean isBlank(String str) {
-        if (str == null || "".equals(str.trim())) {
-            return true;
-        }
-        return false;
-    }
-
-}
