--- src/main/java/com/qm/tool/mail/MailAliUtils.java
@@ -1,4 +1,0 @@
-package com.qm.tool.mail;
-
-public class MailAliUtils {
-}
