+++ src/main/java/com/quanminshangxian/tool/crypto/DesUtils.java
@@ -1,0 +1,86 @@
+package com.quanminshangxian.tool.crypto;
+
+import javax.crypto.Cipher;
+import javax.crypto.SecretKey;
+import javax.crypto.SecretKeyFactory;
+import javax.crypto.spec.DESKeySpec;
+import java.security.Key;
+import java.security.SecureRandom;
+
+/**
+ * DES(Data Encryption Standard) 对称加密/解密
+ * <p>
+ * 数据加密标准算法,和BASE64最明显的区别就是有一个工作密钥，该密钥既用于加密、也用于解密，并且要求密钥是一个长度至少大于8位的字符串
+ */
+public class DesUtils {
+
+    private static final String ALGORITHM_DES = "des";
+    private static SecureRandom secureRandom;
+
+    static {
+        secureRandom = new SecureRandom();
+    }
+
+    /**
+     * DES加密
+     *
+     * @param key     秘钥key
+     * @param content 待加密内容
+     * @return byte[]
+     */
+    public static byte[] DESEncrypt(final String key, final String content) {
+        return processCipher(content.getBytes(), getSecretKey(key), Cipher.ENCRYPT_MODE, ALGORITHM_DES);
+    }
+
+    /**
+     * DES解密
+     *
+     * @param key            秘钥key
+     * @param encoderContent 已加密内容
+     * @return byte[]
+     */
+    public static byte[] DESDecrypt(final String key, final byte[] encoderContent) {
+        return processCipher(encoderContent, getSecretKey(key), Cipher.DECRYPT_MODE, ALGORITHM_DES);
+    }
+
+    /**
+     * 根据key生成秘钥
+     *
+     * @param key 给定key,要求key至少长度为8个字符
+     * @return SecretKey
+     */
+    public static SecretKey getSecretKey(final String key) {
+        try {
+            DESKeySpec desKeySpec = new DESKeySpec(key.getBytes());
+            SecretKeyFactory instance = SecretKeyFactory.getInstance(ALGORITHM_DES);
+            return instance.generateSecret(desKeySpec);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return null;
+    }
+
+    /**
+     * 加密/解密处理
+     *
+     * @param processData 待处理的数据
+     * @param key         提供的密钥
+     * @param opsMode     工作模式
+     * @param algorithm   使用的算法
+     * @return byte[]
+     */
+    private static byte[] processCipher(final byte[] processData, final Key key,
+                                        final int opsMode, final String algorithm) {
+        try {
+            Cipher cipher = Cipher.getInstance(algorithm);
+            //初始化
+            cipher.init(opsMode, key, secureRandom);
+            return cipher.doFinal(processData);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+}
