+++ src/main/java/com/quanminshangxian/tool/crypto/HmacUtils.java
@@ -1,0 +1,36 @@
+package com.quanminshangxian.tool.crypto;
+
+import javax.crypto.Mac;
+import javax.crypto.SecretKey;
+import javax.crypto.spec.SecretKeySpec;
+
+/**
+ * HMAC(Hash Message Authentication Code，散列消息鉴别码)
+ * <p>
+ * 使用一个密钥生成一个固定大小的小数据块，即MAC，并将其加入到消息中，然后传输。接收方利用与发送方共享的密钥进行鉴别认证
+ */
+public class HmacUtils {
+
+    private static final String ALGORITHM_MAC = "HmacMD5";
+    
+    /**
+     * HMAC加密
+     *
+     * @param key     给定秘钥key
+     * @param content 待加密内容
+     * @return String
+     */
+    public static byte[] HMACEncrypt(final String key, final String content) {
+        try {
+            SecretKey secretKey = new SecretKeySpec(key.getBytes(), ALGORITHM_MAC);
+            Mac mac = Mac.getInstance(secretKey.getAlgorithm());
+            //初始化mac
+            mac.init(secretKey);
+            return mac.doFinal(content.getBytes());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+}
