+++ src/main/java/com/quanminshangxian/tool/crypto/Md5Utils.java
@@ -1,0 +1,45 @@
+package com.quanminshangxian.tool.crypto;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+/**
+ * md5加密
+ * <p>
+ * 是一种单向加密算法，只能加密不能解密
+ */
+public class Md5Utils {
+
+    /**
+     * 获取内容的md5加密字符串
+     *
+     * @param content
+     * @return
+     */
+    public static String getMd5(final String content) {
+        try {
+            MessageDigest md = MessageDigest.getInstance("MD5");
+            md.update(content.getBytes());
+            byte b[] = md.digest();
+
+            int i;
+
+            StringBuffer buf = new StringBuffer("");
+            for (int offset = 0; offset < b.length; offset++) {
+                i = b[offset];
+                if (i < 0) {
+                    i += 256;
+                }
+                if (i < 16) {
+                    buf.append("0");
+                }
+                buf.append(Integer.toHexString(i));
+            }
+            return buf.toString();
+        } catch (NoSuchAlgorithmException e) {
+            e.printStackTrace();
+            throw new RuntimeException("md5 encryption failure");
+        }
+    }
+
+}
