+++ src/main/java/com/quanminshangxian/tool/http/HttpUtils.java
@@ -1,0 +1,142 @@
+package com.quanminshangxian.tool.http;
+
+import java.io.*;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.Map;
+
+/**
+ * http 请求工具类
+ */
+public class HttpUtils {
+
+    /**
+     * get请求
+     *
+     * @param path
+     * @param param
+     * @return
+     */
+    public static String get(String path, Map<String, Object> param) {
+        try {
+            if (param != null) {
+                StringBuffer paramBuffer = new StringBuffer();
+                int i = 0;
+                for (String key : param.keySet()) {
+                    if (i == 0) {
+                        paramBuffer.append("?");
+                    } else {
+                        paramBuffer.append("&");
+                    }
+                    paramBuffer.append(key).append("=").append(param.get(key));
+                    i++;
+                }
+                path += paramBuffer;
+            }
+            URL url = new URL(path);    // 把字符串转换为URL请求地址
+            HttpURLConnection connection = (HttpURLConnection) url.openConnection();// 打开连接
+            connection.connect();// 连接会话
+            // 获取输入流
+            BufferedReader br = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+            String line;
+            StringBuilder sb = new StringBuilder();
+            while ((line = br.readLine()) != null) {// 循环读取流
+                sb.append(line);
+            }
+            br.close();// 关闭流
+            connection.disconnect();// 断开连接
+            return sb.toString();
+        } catch (Exception e) {
+            e.printStackTrace();
+            System.out.println("失败!");
+            return null;
+        }
+    }
+
+    /**
+     * post请求
+     *
+     * @param httpUrl
+     * @param param
+     * @return
+     */
+    public static String doPost(String httpUrl, String param) {
+
+        HttpURLConnection connection = null;
+        InputStream is = null;
+        OutputStream os = null;
+        BufferedReader br = null;
+        String result = null;
+        try {
+            URL url = new URL(httpUrl);
+            // 通过远程url连接对象打开连接
+            connection = (HttpURLConnection) url.openConnection();
+            // 设置连接请求方式
+            connection.setRequestMethod("POST");
+            // 设置连接主机服务器超时时间：15000毫秒
+            connection.setConnectTimeout(15000);
+            // 设置读取主机服务器返回数据超时时间：60000毫秒
+            connection.setReadTimeout(60000);
+
+            // 默认值为：false，当向远程服务器传送数据/写数据时，需要设置为true
+            connection.setDoOutput(true);
+            // 默认值为：true，当前向远程服务读取数据时，设置为true，该参数可有可无
+            connection.setDoInput(true);
+            // 设置传入参数的格式:请求参数应该是 name1=value1&name2=value2 的形式。
+            connection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
+            // 设置鉴权信息：Authorization: Bearer da3efcbf-0845-4fe3-8aba-ee040be542c0
+            connection.setRequestProperty("Authorization", "Bearer da3efcbf-0845-4fe3-8aba-ee040be542c0");
+            // 通过连接对象获取一个输出流
+            os = connection.getOutputStream();
+            // 通过输出流对象将参数写出去/传输出去,它是通过字节数组写出的
+            os.write(param.getBytes());
+            // 通过连接对象获取一个输入流，向远程读取
+            if (connection.getResponseCode() == 200) {
+
+                is = connection.getInputStream();
+                // 对输入流对象进行包装:charset根据工作项目组的要求来设置
+                br = new BufferedReader(new InputStreamReader(is, "UTF-8"));
+
+                StringBuffer sbf = new StringBuffer();
+                String temp = null;
+                // 循环遍历一行一行读取数据
+                while ((temp = br.readLine()) != null) {
+                    sbf.append(temp);
+                    sbf.append("\r\n");
+                }
+                result = sbf.toString();
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        } finally {
+            // 关闭资源
+            if (null != br) {
+                try {
+                    br.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (null != os) {
+                try {
+                    os.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+            if (null != is) {
+                try {
+                    is.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+            // 断开与远程地址url的连接
+            connection.disconnect();
+        }
+        return result;
+    }
+
+
+}
