+++ src/main/java/com/quanminshangxian/tool/mail/MailAliUtils.java
@@ -1,0 +1,4 @@
+package com.quanminshangxian.tool.mail;
+
+public class MailAliUtils {
+}
