+++ src/main/java/com/quanminshangxian/tool/mail/MailTxUtils.java
@@ -1,0 +1,4 @@
+package com.quanminshangxian.tool.mail;
+
+public class MailTxUtils {
+}
