--- src/test/java/com/qm/tool/crypto/HmacUtilsTests.java
@@ -1,12 +1,0 @@
-package com.qm.tool.crypto;
-
-import org.junit.Test;
-
-public class HmacUtilsTests {
-
-    @Test
-    public void HMACEncrypt(){
-        byte[] bytes = HmacUtils.HMACEncrypt("4444", "aaaa");
-        System.out.println(new String(bytes));
-    }
-}
