--- src/test/java/com/qm/tool/crypto/Md5UtilsTests.java
@@ -1,12 +1,0 @@
-package com.qm.tool.crypto;
-
-import org.junit.Test;
-
-public class Md5UtilsTests {
-
-    @Test
-    public void getMd5(){
-        String md5 = Md5Utils.getMd5("");
-        System.out.println(md5);
-    }
-}
