+++ src/test/java/com/quanminshangxian/tool/crypto/HmacUtilsTests.java
@@ -1,0 +1,12 @@
+package com.quanminshangxian.tool.crypto;
+
+import org.junit.Test;
+
+public class HmacUtilsTests {
+
+    @Test
+    public void HMACEncrypt(){
+        byte[] bytes = HmacUtils.HMACEncrypt("4444", "aaaa");
+        System.out.println(new String(bytes));
+    }
+}
